+++ lib/domain/business/bloc/favorite_bloc/favorite_list_states.dart
