+++ lib/domain/business/bloc/home_bloc/home_list_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:marvel_app/controllers/heroes_controller.dart';
-import 'package:marvel_app/data/bloc/home_bloc/home_list_events.dart';
-import 'package:marvel_app/data/bloc/home_bloc/home_list_states.dart';
+import 'package:marvel_app/domain/business/bloc/home_bloc/home_list_events.dart';
+import 'package:marvel_app/domain/business/bloc/home_bloc/home_list_states.dart';
 
 class HomeListBloc extends Bloc<HomeListEvent, HomeListStates> {
   final HeroesController _heroesController = HeroesController();
